PF/backend/requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.0.0 
phonenumberslite==8.13.0
Pillow==9.3.0
django-credit-cards==0.4.1
geographiclib==2.0
geopy==2.3.0
python-dateutil==2.8.2
six==1.16.0
django-recurrence==1.11.1
certifi==2022.9.24
django-credit-cards==0.4.1